+++ libs/angular/src/lib/shared/core-control/index.ts
@@ -1,1 +1,2 @@
 export * from './core-control.module'
+export * from './core-control.directive'