+++ libs/core/src/components/datepicker/datepicker.test.ts
@@ -5,14 +5,14 @@ import sinon from 'sinon'
 
 import type { GdsDatepicker } from '@sebgroup/green-core/components/datepicker'
 
+import { GdsButton } from '@sebgroup/green-core/components/button'
+import { GdsDropdown } from '@sebgroup/green-core/components/dropdown'
 import { GdsPopover } from '@sebgroup/green-core/components/popover'
 import {
   getScopedTagName,
   htmlTemplateTagFactory,
 } from '@sebgroup/green-core/scoping'
-import { GdsDropdown } from '@sebgroup/green-core/components/dropdown'
-import { GdsButton } from '@sebgroup/green-core/components/button'
-import { clickOnElement, onlyDate, isWebKit } from '../../utils/testing'
+import { clickOnElement, isWebKit, onlyDate } from '../../utils/testing'
 import { GdsDatePartSpinner } from './date-part-spinner'
 
 import '@sebgroup/green-core/components/datepicker'
@@ -77,9 +77,8 @@ describe('<gds-datepicker>', () => {
       const spinners = el.shadowRoot!.querySelectorAll<GdsDatePartSpinner>(
         getScopedTagName('gds-date-part-spinner'),
       )!
-      const separator = el.shadowRoot!.querySelector<HTMLSpanElement>(
-        '.field .input > span',
-      )!
+      const separator =
+        el.shadowRoot!.querySelector<HTMLSpanElement>('.separator')!
 
       await expect(spinners[0].getAttribute('aria-label')).to.equal('Day')
       await expect(spinners[0].length).to.equal(2)
@@ -235,7 +234,7 @@ describe('<gds-datepicker>', () => {
       const el = await fixture<GdsDatepicker>(
         html`<gds-datepicker label="Date"></gds-datepicker>`,
       )
-      const label = el.shadowRoot!.querySelector<HTMLLabelElement>('#label')!
+      const label = el.shadowRoot!.querySelector<HTMLLabelElement>('label')!
       const spinners = el.shadowRoot!.querySelectorAll<GdsDatePartSpinner>(
         getScopedTagName('gds-date-part-spinner'),
       )!
@@ -518,7 +517,12 @@ describe('<gds-datepicker>', () => {
 
     it('should not overflow the year when trying to increase max year', async () => {
       const el = await fixture<GdsDatepicker>(
-        html`<gds-datepicker value="2034-12-10" min="2014-01-01" max="2034-12-31" open></gds-datepicker>`,
+        html`<gds-datepicker
+          value="2034-12-10"
+          min="2014-01-01"
+          max="2034-12-31"
+          open
+        ></gds-datepicker>`,
       )
       await el.updateComplete
 
@@ -531,19 +535,24 @@ describe('<gds-datepicker>', () => {
       const nextMonthButton = el.shadowRoot!.querySelector<GdsButton>(
         `${getScopedTagName('gds-button')}[aria-label="Next month"]`,
       )!
-      
+
       nextMonthButton.click()
-      
+
       await aTimeout(0)
       await yearDropdown.updateComplete
-      
+
       await expect(monthDropdown.value).to.equal('11')
       await expect(yearDropdown.value).to.equal('2034')
     })
 
     it('should not overflow the year when trying to decrease min year', async () => {
       const el = await fixture<GdsDatepicker>(
-        html`<gds-datepicker value="2014-01-10" min="2014-01-01" max="2034-12-31" open></gds-datepicker>`,
+        html`<gds-datepicker
+          value="2014-01-10"
+          min="2014-01-01"
+          max="2034-12-31"
+          open
+        ></gds-datepicker>`,
       )
       await el.updateComplete
 
@@ -556,15 +565,41 @@ describe('<gds-datepicker>', () => {
       const prevMonthButton = el.shadowRoot!.querySelector<GdsButton>(
         `${getScopedTagName('gds-button')}[aria-label="Previous month"]`,
       )!
-      
+
       prevMonthButton.click()
-      
+
       await aTimeout(0)
       await monthDropdown.updateComplete
       await yearDropdown.updateComplete
-      
+
       await expect(monthDropdown.value).to.equal('0')
       await expect(yearDropdown.value).to.equal('2014')
+    })
+  })
+
+  describe('Accessibility', () => {
+    it('should pass axe smoketest', async () => {
+      const el = await fixture<GdsDatepicker>(
+        html`<gds-datepicker
+          value="2014-01-10"
+          min="2014-01-01"
+          max="2034-12-31"
+          open
+        ></gds-datepicker>`,
+      )
+      await el.updateComplete
+      await expect(el).to.be.accessible({
+        ignoredRules: ['color-contrast'],
+      })
+    })
+
+    it('should have a label for #spinner-0', async () => {
+      const el = await fixture<GdsDatepicker>(
+        html`<gds-datepicker label="Date"></gds-datepicker>`,
+      )
+      const label =
+        el.shadowRoot!.querySelector<HTMLLabelElement>('[for="spinner-0"]')!
+      expect(label).to.exist
     })
   })
 })