+++ libs/core/src/components/dropdown/dropdown.styles.ts
@@ -7,78 +7,36 @@ const style = css`
     :host {
       display: flex;
       flex-direction: column;
-      color: var(--gds-color-l3-content-tertiary);
-      font-family: inherit;
+      width: 100%;
+      gap: var(--gds-space-xs);
+      contain: layout;
+      isolation: isolate;
+      --_transition: all 368ms cubic-bezier(0.4, 0, 0.2, 1);
+    }
+
+    #field::part(_base) {
+      cursor: pointer;
     }
 
     button {
-      display: flex;
-      align-items: center;
-      justify-content: space-between;
-      gap: var(--gds-space-m);
-      height: var(--gds-space-3xl);
-      border-style: solid;
-      border-width: var(--gds-space-4xs);
-      border-radius: var(--gds-space-xs);
-      padding-inline: var(--gds-space-m);
-      cursor: pointer;
-      box-sizing: border-box;
-      font-size: var(--gds-space-m);
+      appearance: none;
+      background: transparent;
+      border-width: 0;
       font-family: inherit;
-      border-color: var(--gds-color-l3-border-secondary);
-      color: var(--gds-color-l3-content-tertiary);
-      background-color: var(--gds-color-l3-background-secondary);
-      outline-color: transparent;
-      outline-offset: var(--gds-space-3xs);
-      outline-style: solid;
-      outline-width: var(--gds-space-3xs);
-      transition:
-        background-color 0.2s,
-        border-color 0.2s,
-        color 0.2s;
+      font-size: inherit;
+      line-height: inherit;
+      color: currentColor;
+      cursor: pointer;
+      padding: 0;
+      text-align: left;
+      flex: 0 1 100%;
 
-      &:focus {
-        outline-color: color-mix(in srgb, currentcolor, #000 100%);
+      /* This prevents overflow when the text is too long */
+      display: grid;
+    }
 
-        &:not(:focus-visible) {
-          outline-color: transparent;
-        }
-      }
-
-      &.small {
-        height: var(--gds-space-xl);
-        padding: var(--gds-space-s);
-        padding-right: var(--gds-space-xs);
-        font-size: calc(var(--gds-space-s) + 2px);
-
-        .icon {
-          width: var(--gds-space-l);
-          height: var(--gds-space-l);
-        }
-      }
-
-      @media (pointer: fine) {
-        &:hover {
-          background-color: color-mix(
-            in srgb,
-            var(--gds-color-l3-background-secondary),
-            var(--gds-color-l3-states-light-hover)
-          );
-        }
-      }
-
-      .icon {
-        display: flex;
-        align-items: center;
-        justify-content: center;
-        box-sizing: border-box;
-        height: var(--gds-space-xl);
-        width: var(--gds-space-xl);
-      }
-
-      &[aria-expanded='true'] ::part(icon) {
-        transform: scaleY(-1);
-      }
+    button:focus {
+      outline: none;
     }
 
     slot[name='trigger'] > span {
@@ -86,14 +44,6 @@ const style = css`
       overflow: hidden;
       text-overflow: ellipsis;
       white-space: nowrap;
-    }
-
-    label {
-      font-size: var(--gds-text-size-detail-m);
-      line-height: var(--gds-text-line-height-detail-m);
-      font-weight: var(--gds-text-weight-book);
-      padding-block: var(--gds-space-2xs);
-      font-family: inherit;
     }
 
     input[type='text'] {
@@ -125,6 +75,10 @@ const style = css`
         font-weight: var(--gds-text-weight-book);
         color: var(--gds-color-l3-content-disabled);
       }
+    }
+
+    [aria-expanded='true'] [slot='trail'] {
+      transform: scaleY(-1);
     }
   }
 `