+++ libs/chlorophyll/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 3.6.13
+
+### Patch Changes
+
+- 3331d19: Tabs not selected now have same size as selected tabs
+
 ## 3.6.12
 
 ### Patch Changes